requirements.txt
aiogram==3.22.0
aiohttp==3.12.15
aiofiles==24.1.0
aiosqlite==0.21.0
SQLAlchemy==2.0.43
pydantic==2.11.7
typing-extensions==4.15.0
python-dotenv